Prevent o2m-records to be deleted on client-side
When I delete some records of an one2many-field in an on-change-Event, the
deleted records are sent to the client and the client hides them in the list.
Pressing the save button the record loads again including the deleted o2m-records.

The reason is the force_remove-flag. This will delete the record on the
client-side and therefore this change will not be reflected on the server side.

issue3091
review744002
(transplanted from e40190b28b312f3dd63bceb167698a885826ea0e)