reset tree_states_done when setting a screen group
The O2M screen's tree_states_done property is filled with the view id, this
property is used to know when the state must be set or not. This has the effect
of not restoring the state of O2M widget when switching from a record to
another. This change fixes this behavior.

issue3609
review2791003