Don't set cursor when displaying widget and correctly manage selection in list view
When clicking on new, a new record instance is set as current_record and display
is called. But as the changeset ff6dc4d5bff6 adds a call to display with
set_cursor, it triggers the focus out of the widget having the cursor which
could set the value of the previous record to the new record if the widget was
not yet displayed.
So display on widget should never set the cursor but indeed it is the display of
the list view that must set the correct selection when the store is changed.

issue3757
review4481002