Let GenericText instantiate renderer for Selection
The GenericText connect the renderer to the send_keys callback which manage the
keypress event for editable tree. It will be lost if the Selection widget
override the renderer. GenericText was designed to allow to pass the renderer
class to instantiate so Selection widget should use this method.

issue5044
review15741002