Use richtext widget for translation
The Toolbar is linked to a TextView so one Toolbar can be created per
translation TextView.
The API of TranslateMixin is modified into instance method to allow to call
helpers from the widget.
The TextBox is reworked to ease its usage by the RichTextBox.
The language label of the TranslateDialog are always vertically aligned to not
be in front of the toolbar.

issue6891
review38951003